FIX: r8962 Ticket #2116 - --object_id argument change type
git-svn-id: file:///data/svn/enum/server/trunk@8986 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ src/cli_admin/objectclient.cc
@@ -89,7 +89,7 @@ void
 ObjectClient::new_state_request()
 {
     callHelp(m_conf, no_help);
-    Register::TID id = m_conf.get<unsigned int>(OBJECT_ID_NAME);
+    Register::TID id = m_conf.get<unsigned long long>(OBJECT_ID_NAME);
     unsigned int state = m_conf.get<unsigned int>(OBJECT_NEW_STATE_REQUEST_NAME);
     int res = createObjectStateRequest(
             id, state
